refactor(database): extract credit bucket scope helpers
Move buildCreditBucketScopeWhere and buildCreditBucketScopeSql into a
shared helper module for Phase 5b enterprise pool gating.
